fix логаут венул как был и убрал редирект на /auth/login чтоб не ломало регистрацию
